Move <PaneItem> to a new lib/atom/ directory
