+++ lib/atom/pane-item.js
