+++ lib/controllers/root-controller.js
@@ -9,7 +9,7 @@ import {CompositeDisposable} from 'event-kit';
 
 import StatusBar from '../views/status-bar';
 import Panel from '../views/panel';
-import PaneItem from '../views/pane-item';
+import PaneItem from '../atom/pane-item';
 import DockItem from '../views/dock-item';
 import CloneDialog from '../views/clone-dialog';
 import OpenIssueishDialog from '../views/open-issueish-dialog';
